fix(lazy): migrate to lazy properly
I think most of the leg-work should be done with the migration, please
do look through it though.
