--- .config/nvim/after/plugin/comment.lua
@@ -1,3 +1,0 @@
-require("Comment").setup({
-	pre_hook = require("ts_context_commentstring.integrations.comment_nvim").create_pre_hook(),
-})
