--- .config/nvim/after/plugin/copilot.lua
@@ -1,9 +1,0 @@
-require("copilot").setup({
-	suggestion = {
-		auto_trigger = true,
-		keymap = {
-			accept = "<C-/>",
-			dismiss = "<M-/>",
-		},
-	},
-})
