--- .config/nvim/after/plugin/dressing.lua
@@ -1,1 +1,0 @@
-require("dressing").setup()
