--- .config/nvim/after/plugin/fidget.lua
@@ -1,6 +1,0 @@
--- Turn on LSP, formatting, and linting status and progress information
-require("fidget").setup({
-  text = {
-    spinner = "dots_negative",
-  },
-})
