--- .config/nvim/after/plugin/illuminate.lua
@@ -1,13 +1,0 @@
-require("illuminate").configure({
-	under_cursor = false,
-	filetypes_denylist = {
-		"DressingSelect",
-		"Outline",
-		"TelescopePrompt",
-		"alpha",
-		"harpoon",
-		"toggleterm",
-		"neo-tree",
-		"Spectre",
-	},
-})
