--- .config/nvim/after/plugin/indent_blankline.lua
@@ -1,7 +1,0 @@
-require("ibl").setup({
-	scope = {
-		enabled = false,
-		show_start = false,
-		show_end = false,
-	},
-})
