--- .config/nvim/after/plugin/lualine.lua
@@ -1,65 +1,0 @@
-local harpoon = require("harpoon.mark")
-local theme = require("catppuccin.palettes").get_palette("macchiato")
-
-local function truncate_branch_name(branch)
-	if not branch or branch == "" then
-		return ""
-	end
-
-	-- Match the branch name to the specified format
-	local _, _, ticket_number = string.find(branch, "skdillon/sko%-(%d+)%-")
-
-	-- If the branch name matches the format, display sko-{ticket_number}, otherwise display the full branch name
-	if ticket_number then
-		return "sko-" .. ticket_number
-	else
-		return branch
-	end
-end
-
-local function harpoon_component()
-	local total_marks = harpoon.get_length()
-
-	if total_marks == 0 then
-		return ""
-	end
-
-	local current_mark = "—"
-
-	local mark_idx = harpoon.get_current_index()
-	if mark_idx ~= nil then
-		current_mark = tostring(mark_idx)
-	end
-
-	return string.format("󱡅 %s/%d", current_mark, total_marks)
-end
-
-local function kitty_component()
-	if vim.g.kitty_layout_is_stack then
-		return " "
-	end
-
-	return ""
-end
-
-require("lualine").setup({
-	options = {
-		theme = "catppuccin",
-		globalstatus = true,
-	},
-	sections = {
-		lualine_b = {
-			{ "branch", icon = "", fmt = truncate_branch_name },
-			harpoon_component,
-			"diff",
-			"diagnostics",
-		},
-		lualine_c = {
-			{ "filename", path = 1 },
-		},
-		lualine_x = {
-			kitty_component,
-			"filetype",
-		},
-	},
-})
