--- .config/nvim/after/plugin/telescope.lua
@@ -1,23 +1,0 @@
-local actions = require("telescope.actions")
-
-require("telescope").setup({
-	defaults = {
-		mappings = {
-			i = {
-				["<C-k>"] = actions.move_selection_previous,
-				["<C-j>"] = actions.move_selection_next,
-				["<C-q>"] = actions.send_selected_to_qflist + actions.open_qflist,
-				["<C-x>"] = actions.delete_buffer,
-			},
-		},
-		file_ignore_patterns = {
-			"node_modules",
-			"yarn.lock",
-			".git",
-		},
-		hidden = true,
-	},
-})
-
--- Enable telescope fzf native, if installed
-pcall(require("telescope").load_extension, "fzf")
