--- .config/nvim/after/plugin/tsc.lua
@@ -1,1 +1,0 @@
-require("tsc").setup()
