--- .config/nvim/after/plugin/ufo.lua
@@ -1,5 +1,0 @@
-require("ufo").setup({
-	provider_selector = function(_bufnr, _filetype, _buftype)
-		return { "treesitter", "indent" }
-	end,
-})
