+++ .config/nvim/lua/plugins/autopairs_and_cmp.lua
@@ -1,0 +1,26 @@
+return {
+	{
+		"hrsh7th/nvim-cmp",
+		event = { "BufReadPost", "BufNewFile" },
+		dependencies = {
+			"hrsh7th/cmp-nvim-lsp",
+			"hrsh7th/cmp-buffer",
+			"hrsh7th/cmp-path",
+			"L3MON4D3/LuaSnip",
+			"saadparwaiz1/cmp_luasnip",
+			"rafamadriz/friendly-snippets",
+			"onsails/lspkind.nvim",
+			"windwp/nvim-ts-autotag",
+			"windwp/nvim-autopairs",
+		},
+		config = function()
+			local cmp_autopairs = require("nvim-autopairs.completion.cmp")
+			local cmp = require("cmp")
+
+			require("nvim-autopairs").setup()
+
+			-- Integrate nvim-autopairs with cmp
+			cmp.event:on("confirm_done", cmp_autopairs.on_confirm_done())
+		end,
+	},
+}
