+++ .config/nvim/lua/plugins/copilot.lua
@@ -1,0 +1,17 @@
+return {
+	{
+		"zbirenbaum/copilot.lua",
+		event = { "BufEnter" },
+		config = function()
+			require("copilot").setup({
+				suggestion = {
+					auto_trigger = true,
+					keymap = {
+						accept = "<C-/>",
+						dismiss = "<M-/>",
+					},
+				},
+			})
+		end,
+	},
+}
