+++ .config/nvim/lua/plugins/dressing.lua
@@ -1,0 +1,8 @@
+return {
+	{
+		"stevearc/dressing.nvim",
+		config = function()
+			require("dressing").setup()
+		end,
+	},
+}
