+++ .config/nvim/lua/plugins/fidget.lua
@@ -1,0 +1,15 @@
+return {
+	{
+		"j-hui/fidget.nvim",
+		tag = "legacy",
+		event = { "BufEnter" },
+		config = function()
+			-- Turn on LSP, formatting, and linting status and progress information
+			require("fidget").setup({
+				text = {
+					spinner = "dots_negative",
+				},
+			})
+		end,
+	},
+}
