+++ .config/nvim/lua/plugins/gitsigns.lua
@@ -1,0 +1,9 @@
+return {
+	{
+		"lewis6991/gitsigns.nvim",
+		event = "VeryLazy",
+		config = function()
+			require("gitsigns").setup()
+		end,
+	},
+}
