+++ .config/nvim/lua/plugins/indent_blankline.lua
@@ -1,0 +1,13 @@
+return {
+	{
+		"lukas-reineke/indent-blankline.nvim",
+		event = "BufEnter",
+		opts = {
+			scope = {
+				enabled = false,
+				show_start = false,
+				show_end = false,
+			},
+		},
+	},
+}
