+++ .config/nvim/lua/plugins/tsc.lua
@@ -1,0 +1,9 @@
+return {
+	{
+		"dmmulroy/tsc.nvim",
+		ft = { "typescript", "typescriptreact" },
+		config = function()
+			require("tsc").setup()
+		end,
+	},
+}
