+++ .config/nvim/lua/plugins/ufo.lua
@@ -1,0 +1,18 @@
+return {
+
+	{
+		"kevinhwang91/nvim-ufo",
+		event = "BufEnter",
+		dependencies = {
+			"kevinhwang91/promise-async",
+		},
+		config = function()
+			--- @diagnostic disable: unused-local
+			require("ufo").setup({
+				provider_selector = function(_bufnr, _filetype, _buftype)
+					return { "treesitter", "indent" }
+				end,
+			})
+		end,
+	},
+}
